Remove the ability to change the iptables comments prefix
The iptables comments prefix will be used during the uninstallation of
the transparent proxy. Allowing this prefix to be configurable adds
unnecessary complexity, as it would require an additional flag during the
uninstallation process. This could potentially lead to issues if the
prefix is changed during installation but not specified during
uninstallation. Removing this configurability simplifies the process and
reduces the risk of errors.

Signed-off-by: Bart Smykla <bartek@smykla.com>